fix(cli): fix dev build and move docs index to static generation
- Fix dev.mjs: esbuild context() was never calling rebuild/watch,
  changed to esbuild.build() so dev builds actually produce output
- Add docs/scripts/generate-docs-index.ts: standalone script that
  reads MDX frontmatter to generate public/__docs__/index.json
- Remove docs/app/__docs__/index.json/route.ts: replaced dynamic
  Next.js route with static file served from public/
- Integrate into docs generate:all pipeline

The docs command was returning 404 in dev mode because:
1. dev.mjs created an esbuild context but never built anything
2. __docs__/index.json was a Next.js route handler only available
   at build time, not during development
