+++ packages/cli/dev.mjs
@@ -1,20 +1,22 @@
-import { context } from "esbuild";
+import esbuild from "esbuild";
 
 import pkg from "./package.json" with { type: "json" };
 
-context({
-  entryPoints: ["./src/index.ts"],
-  outfile: "./bin/index.mjs",
-  bundle: true,
-  write: true,
-  treeShaking: true,
-  sourcemap: false,
-  minify: true,
-  format: "esm",
-  platform: "node",
-  target: ["esnext"],
-  external: [...Object.keys(pkg.dependencies), "__temp.mjs"],
-  define: {
-    "process.env.NODE_ENV": `"dev"`,
-  },
-}).catch(() => process.exit(1));
+esbuild
+  .build({
+    entryPoints: ["./src/index.ts"],
+    outfile: "./bin/index.mjs",
+    bundle: true,
+    write: true,
+    treeShaking: true,
+    sourcemap: false,
+    minify: true,
+    format: "esm",
+    platform: "node",
+    target: ["esnext"],
+    external: [...Object.keys(pkg.dependencies), "__temp.mjs"],
+    define: {
+      "process.env.NODE_ENV": `"dev"`,
+    },
+  })
+  .catch(() => process.exit(1));
